backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
